requirements.txt
Django==5.2.9
djangorestframework>=3.12
psycopg2-binary>=2.9
